+++ src/professores/professores.service.ts
@@ -1,5 +1,5 @@
 /* eslint-disable prettier/prettier */
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { Prisma } from '@prisma/client';
 import { PrismaService } from 'src/prisma.service';
 
@@ -8,27 +8,60 @@ export class ProfessoresService {
   constructor(private prisma: PrismaService){}
 
   async create(createProfessoreDto: Prisma.ProfessorCreateInput) {
+    
+    const matricula = createProfessoreDto.matricula;
+    
+    const professor = await this.prisma.professor.findUnique({
+      where : {matricula}
+    })
+
+    if(professor){
+      throw new HttpException('Professor já cadastrado', HttpStatus.BAD_REQUEST)
+    }
+    
     return this.prisma.professor.create({
       data: createProfessoreDto
     })
   }
 
-  async findOne(matricula: number) {
-    return this.prisma.professor.findUnique({
-      where:{matricula}
+  async findOne(where: Prisma.ProfessorWhereUniqueInput) {
+
+    const professor = await this.prisma.professor.findUnique({
+      where
+    })
+    if(!professor){
+      throw new HttpException('Professor não encontrado', HttpStatus.BAD_REQUEST)
+    } else {
+      return professor
+    }
+
+  }
+
+  async update(where: Prisma.ProfessorWhereUniqueInput, data: Prisma.ProfessorUpdateInput) {
+    const professor = await this.prisma.professor.findUnique({
+      where
+    })
+    if(!professor){
+      throw new HttpException('Professor não encontrado', HttpStatus.BAD_REQUEST)
+    }
+
+    return  this.prisma.professor.update({
+      where,
+      data
     })
   }
 
-  async update(matricula: number, updateProfessoreDto: Prisma.ProfessorUpdateInput) {
-    return this.prisma.professor.update({
-      where: {matricula},
-      data: updateProfessoreDto
+  async remove(where: Prisma.ProfessorWhereUniqueInput) {
+    
+    const professor = await this.prisma.professor.findUnique({
+      where
     })
-  }
-
-  async remove(matricula: number) {
+    if(!professor){
+      throw new HttpException('Professor não encontrado', HttpStatus.BAD_REQUEST)
+    }
+    
     return this.prisma.professor.delete({
-      where: {matricula}
+      where
     })
   }
 }
